+++ packages/yoastseo/src/languageProcessing/researches/getLongCenterAlignedTexts.js
